conf: cleanup and relocation of common parsing functions and helpers
 - legacy.[ch]: strictly legacy .conf parser boilerplate only
 - conf.[ch]: .conf parser and generic configuration functionality
